+++ perl/FSM/Composition/Plan.pm
@@ -16,6 +16,7 @@ sub new ($class, %args) {
         nets => $args{nets} || [],
         instances => $args{instances} || [],
         auxiliary_assignments => $args{auxiliary_assignments} || [],
+        shared_datapath_candidates => $args{shared_datapath_candidates} || [],
         raw_spec => $args{raw_spec},
     }, $class;
 }
@@ -28,6 +29,7 @@ sub resolved_links ($self) { return $self->{resolved_links} }
 sub nets ($self) { return $self->{nets} }
 sub instances ($self) { return $self->{instances} }
 sub auxiliary_assignments ($self) { return $self->{auxiliary_assignments} }
+sub shared_datapath_candidates ($self) { return $self->{shared_datapath_candidates} }
 sub raw_spec ($self) { return $self->{raw_spec} }
 
 1;