+++ t/139-composition-shared-datapath-candidate-metadata.t
@@ -78,6 +78,7 @@ FSM
                 width => 8,
                 interface_type => 'data',
                 storage_class => 'registered',
+                reset_value => "8'h00",
                 contributor_count => 2,
                 contributors => [
                     {
@@ -88,6 +89,7 @@ FSM
                         drive_intent => {
                             multiplexer_type => 'flop',
                             default_value => 'status_bus',
+                            reset_value => "8'h00",
                             driver_count => 1,
                             driver_blocks => ['-state0'],
                             rhs_values => ["8'd1"],
@@ -111,6 +113,7 @@ FSM
                         drive_intent => {
                             multiplexer_type => 'flop',
                             default_value => 'status_bus',
+                            reset_value => "8'h00",
                             driver_count => 1,
                             driver_blocks => ['-state0'],
                             rhs_values => ["8'd2"],